% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{center_points}
\alias{center_points}
\title{Center points (subtract column means)}
\usage{
center_points(values)
}
\arguments{
\item{values}{Numeric matrix}
}
\description{
Center points (subtract column means)
}
