% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choose_two}
\alias{choose_two}
\title{Number of upper-triangle pairs for n codes}
\usage{
choose_two(n)
}
\arguments{
\item{n}{Number of codes}
}
\value{
Integer: n*(n-1)/2
}
\description{
Number of upper-triangle pairs for n codes
}
