% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_connections}
\alias{code_connections}
\title{Pairwise products → upper-triangle vector}
\usage{
code_connections(v)
}
\arguments{
\item{v}{Numeric vector of code values}
}
\description{
Pairwise products → upper-triangle vector
}
