% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connection_matrix}
\alias{connection_matrix}
\title{Core connection matrix for one ground+response pair}
\usage{
connection_matrix(ground, response, response_weight = 1, ordered = TRUE)
}
\arguments{
\item{ground}{Numeric row vector of ground (context) code values}

\item{response}{Numeric row vector of response code values}

\item{response_weight}{Scalar weight applied to the response self-connection}

\item{ordered}{TRUE = directed; FALSE = undirected}
}
\description{
Core connection matrix for one ground+response pair
}
