% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connection_names}
\alias{connection_names}
\title{Code-name pairs for upper-triangle positions ("A & B")}
\usage{
connection_names(v)
}
\arguments{
\item{v}{Character vector of code names}
}
\description{
Code-name pairs for upper-triangle positions ("A & B")
}
