% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deflate}
\alias{deflate}
\title{Project a matrix onto the hyperplane orthogonal to a unit-norm axis}
\usage{
deflate(data, axis)
}
\arguments{
\item{data}{Numeric matrix (n_units x n_dims)}

\item{axis}{Numeric vector of length n_dims, unit-norm}
}
\value{
Numeric matrix of the same shape as \code{data}
}
\description{
Computes \code{data - (data \%*\% axis) \%*\% t(axis)}. The caller is
responsible for ensuring \code{axis} is unit-norm.
}
