% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{directed_node_positions_combine_pairs}
\alias{directed_node_positions_combine_pairs}
\title{Directed node positions with paired ground+response rows combined}
\usage{
directed_node_positions_combine_pairs(line_weights, points, num_dims)
}
\arguments{
\item{line_weights}{Numeric matrix (units x connections)}

\item{points}{Numeric matrix of rotated points (units x dims)}

\item{num_dims}{Number of dimensions}
}
\value{
List with nodes, centroids, weights, points
}
\description{
Directed node positions with paired ground+response rows combined
}
