% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_dist_correlation}
\alias{dist_dist_correlation}
\title{Distance-distance matrix correlation for stability analysis}
\usage{
dist_dist_correlation(X, Y)
}
\arguments{
\item{X}{Numeric matrix}

\item{Y}{Numeric matrix}
}
\description{
Distance-distance matrix correlation for stability analysis
}
