% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{door_ema}
\alias{door_ema}
\title{Door EMA smoothing across units}
\usage{
door_ema(conn_counts, unit_row_indices, alpha = 0.1, segment_ids = integer())
}
\arguments{
\item{conn_counts}{Numeric matrix of connection counts}

\item{unit_row_indices}{List of integer vectors of row indices}

\item{alpha}{Smoothing factor}

\item{segment_ids}{Optional integer vector of segment IDs}
}
\description{
Door EMA smoothing across units
}
