% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{door_ema_block}
\alias{door_ema_block}
\title{Door EMA smoothing for a single block}
\usage{
door_ema_block(block, alpha = 0.1, segment_ids = integer())
}
\arguments{
\item{block}{Numeric matrix of connection counts}

\item{alpha}{Smoothing factor in (0, 1]}

\item{segment_ids}{Integer vector of segment IDs}
}
\description{
Door EMA smoothing for a single block
}
