% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ena_correlation}
\alias{ena_correlation}
\title{Pearson correlation with CI between ENA points and centroids}
\usage{
ena_correlation(points, centroids, conf_level = 0.95)
}
\arguments{
\item{points}{Numeric matrix (units x dims)}

\item{centroids}{Numeric matrix (units x dims)}

\item{conf_level}{Confidence level (default 0.95)}
}
\description{
Pearson correlation with CI between ENA points and centroids
}
