% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_trajectory_curve}
\alias{eval_trajectory_curve}
\title{Evaluate 2D parametric polynomial curve at specified time points}
\usage{
eval_trajectory_curve(coeffs_x, coeffs_y, t_eval)
}
\arguments{
\item{coeffs_x}{Polynomial coefficients for X}

\item{coeffs_y}{Polynomial coefficients for Y}

\item{t_eval}{Numeric vector of time evaluation points}
}
\description{
Evaluate 2D parametric polynomial curve at specified time points
}
