% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_trajectory_derivatives}
\alias{eval_trajectory_derivatives}
\title{Evaluate trajectory differential geometry (velocities, speed, curvature, turns)}
\usage{
eval_trajectory_derivatives(coeffs_x, coeffs_y, t_eval)
}
\arguments{
\item{coeffs_x}{Polynomial coefficients for X}

\item{coeffs_y}{Polynomial coefficients for Y}

\item{t_eval}{Numeric vector of time evaluation points}
}
\description{
Evaluate trajectory differential geometry (velocities, speed, curvature, turns)
}
