% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fold_directed_network}
\alias{fold_directed_network}
\title{Fold a directed (n*n) vector into an undirected upper-triangle vector}
\usage{
fold_directed_network(v)
}
\arguments{
\item{v}{Numeric vector of length n*n}
}
\description{
Fold a directed (n*n) vector into an undirected upper-triangle vector
}
