% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integrated_trajectory_distance}
\alias{integrated_trajectory_distance}
\title{Integrated Euclidean distance between two trajectory polynomial curves}
\usage{
integrated_trajectory_distance(
  coeffs_ax,
  coeffs_ay,
  coeffs_bx,
  coeffs_by,
  t_start = 0,
  t_end = 1
)
}
\arguments{
\item{coeffs_ax}{Polynomial coeffs for Curve A X}

\item{coeffs_ay}{Polynomial coeffs for Curve A Y}

\item{coeffs_bx}{Polynomial coeffs for Curve B X}

\item{coeffs_by}{Polynomial coeffs for Curve B Y}

\item{t_start}{Integration start (default 0.0)}

\item{t_end}{Integration end (default 1.0)}
}
\description{
Integrated Euclidean distance between two trajectory polynomial curves
}
