% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lagged_trajectory_distance}
\alias{lagged_trajectory_distance}
\title{Lagged curve distance between follower and leader}
\usage{
lagged_trajectory_distance(
  coeffs_fol_x,
  coeffs_fol_y,
  coeffs_ldr_x,
  coeffs_ldr_y,
  lag = 0
)
}
\arguments{
\item{coeffs_fol_x}{Polynomial coeffs for Follower X}

\item{coeffs_fol_y}{Polynomial coeffs for Follower Y}

\item{coeffs_ldr_x}{Polynomial coeffs for Leader X}

\item{coeffs_ldr_y}{Polynomial coeffs for Leader Y}

\item{lag}{Time lag in [0, 1)}
}
\description{
Lagged curve distance between follower and leader
}
