% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_center_data}
\alias{lq_center_data}
\title{Center data (subtract column means)}
\usage{
lq_center_data(values)
}
\arguments{
\item{values}{Numeric matrix}
}
\description{
Center data (subtract column means)
}
