% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_directed_node_positions}
\alias{lq_directed_node_positions}
\title{Least-squares node positions for directed ENA}
\usage{
lq_directed_node_positions(line_weights, points, num_dims)
}
\arguments{
\item{line_weights}{Numeric matrix (units x connections)}

\item{points}{Numeric matrix of rotated points (units x dims)}

\item{num_dims}{Number of dimensions}
}
\value{
List with nodes, centroids, weights, points
}
\description{
Least-squares node positions for directed ENA
}
