% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_directed_to_upper_tri}
\alias{lq_directed_to_upper_tri}
\title{Fold a directed (n*n) vector into an undirected upper-triangle vector}
\usage{
lq_directed_to_upper_tri(v)
}
\arguments{
\item{v}{Numeric vector of length n*n}
}
\description{
Fold a directed (n*n) vector into an undirected upper-triangle vector
}
