% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_skip_sphere_norm}
\alias{lq_skip_sphere_norm}
\title{Max-norm scaling (divide all rows by the largest row L2 norm)}
\usage{
lq_skip_sphere_norm(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\description{
Max-norm scaling (divide all rows by the largest row L2 norm)
}
