% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_svector_to_upper_tri}
\alias{lq_svector_to_upper_tri}
\title{Code-name pairs for upper-triangle positions ("A & B")}
\usage{
lq_svector_to_upper_tri(v)
}
\arguments{
\item{v}{Character vector of code names}
}
\description{
Code-name pairs for upper-triangle positions ("A & B")
}
