% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_tri_indices}
\alias{lq_tri_indices}
\title{Upper-triangle index pairs}
\usage{
lq_tri_indices(len, row = -1L)
}
\arguments{
\item{len}{Number of codes (side length of square matrix)}

\item{row}{-1 = both rows, 0 = row indices only, 1 = col indices only}
}
\description{
Upper-triangle index pairs
}
