% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lq_vector_to_upper_tri}
\alias{lq_vector_to_upper_tri}
\title{Pairwise products → upper-triangle vector}
\usage{
lq_vector_to_upper_tri(v)
}
\arguments{
\item{v}{Numeric vector of code values}
}
\description{
Pairwise products → upper-triangle vector
}
