% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{network_to_vector}
\alias{network_to_vector}
\title{Flatten an adjacency matrix to a connection vector}
\usage{
network_to_vector(x, full = TRUE)
}
\arguments{
\item{x}{Numeric matrix}

\item{full}{TRUE = full n*n (directed); FALSE = upper triangle (undirected)}
}
\description{
Flatten an adjacency matrix to a connection vector
}
