% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{node_positions}
\alias{node_positions}
\title{Least-squares node positions for undirected ENA}
\usage{
node_positions(adj_mats, t, num_dims)
}
\arguments{
\item{adj_mats}{Numeric matrix of line weights (units x connections)}

\item{t}{Numeric matrix of rotated points (units x dims)}

\item{num_dims}{Number of dimensions}
}
\value{
List with nodes, centroids, weights, points
}
\description{
Least-squares node positions for undirected ENA
}
