% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_networks}
\alias{normalize_networks}
\title{Row-wise L2 (sphere) normalization}
\usage{
normalize_networks(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\description{
Row-wise L2 (sphere) normalization
}
