% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_trajectory_distance}
\alias{pairwise_trajectory_distance}
\title{Pairwise integrated trajectory distance matrix}
\usage{
pairwise_trajectory_distance(all_coeffs_x, all_coeffs_y)
}
\arguments{
\item{all_coeffs_x}{List of X coefficient vectors}

\item{all_coeffs_y}{List of Y coefficient vectors}
}
\description{
Pairwise integrated trajectory distance matrix
}
