% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_networks}
\alias{scale_networks}
\title{Max-norm scaling (divide all rows by the largest row L2 norm)}
\usage{
scale_networks(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\description{
Max-norm scaling (divide all rows by the largest row L2 norm)
}
