% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{signed_turn_lag}
\alias{signed_turn_lag}
\title{Signed turn-lag distance between two agents at specific lag delta}
\usage{
signed_turn_lag(pts_a, pts_b, times_a, times_b, delta)
}
\arguments{
\item{pts_a}{Points matrix for agent A (n_a x 2)}

\item{pts_b}{Points matrix for agent B (n_b x 2)}

\item{times_a}{Turn times for agent A}

\item{times_b}{Turn times for agent B}

\item{delta}{Signed lag in turns}
}
\description{
Signed turn-lag distance between two agents at specific lag delta
}
