% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sweep_signed_turn_lags}
\alias{sweep_signed_turn_lags}
\title{Sweep signed turn lags to find optimal leader-follower lag}
\usage{
sweep_signed_turn_lags(pts_a, pts_b, times_a, times_b, max_lag = 15L)
}
\arguments{
\item{pts_a}{Points matrix for agent A (n_a x 2)}

\item{pts_b}{Points matrix for agent B (n_b x 2)}

\item{times_a}{Turn times for agent A}

\item{times_b}{Turn times for agent B}

\item{max_lag}{Maximum turn lag to evaluate (default 15)}
}
\description{
Sweep signed turn lags to find optimal leader-follower lag
}
