% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libqe-package.R
\docType{package}
\name{libqe-package}
\alias{libqe-package}
\alias{libqe}
\title{libqe: Shared C++ Core for Quantitative Ethnography Packages}
\description{
Header-only C++ library providing shared computational primitives for
rENA, tma, and other Quantitative Ethnography packages. Exposes four
modules via Rcpp-wrapped \code{lq_*} functions:
}
\details{
\describe{
  \item{Adjacency}{Upper-triangle index pairs and vector/matrix conversions.
    See \code{\link{lq_tri_indices}}, \code{\link{lq_vector_to_upper_tri}},
    \code{\link{lq_svector_to_upper_tri}}.}
  \item{Normalization}{Row-wise L2 sphere norm and max-norm scaling.
    See \code{\link{lq_sphere_norm}}, \code{\link{lq_skip_sphere_norm}}.}
  \item{Modeling}{Column centering, ENA correlation, and least-squares node
    positions (undirected and directed).
    See \code{\link{lq_center_data}}, \code{\link{lq_ena_correlation}},
    \code{\link{lq_lws_lsq_positions}}, \code{\link{lq_directed_node_positions}}.}
  \item{Accumulation}{Stanza-window (rENA) and ground/response/tensor (tma)
    accumulation primitives.
    See \code{\link{lq_stanza_window}}, \code{\link{lq_accumulate_unit}},
    \code{\link{lq_apply_tensor}}.}
}

\strong{Using libqe headers in your own R package:}

Add to \file{DESCRIPTION}:
\preformatted{
  LinkingTo: Rcpp, RcppArmadillo, libqe
}

Then in your \file{.cpp} source:
\preformatted{
  // [[Rcpp::depends(RcppArmadillo, libqe)]]
  #include <RcppArmadillo.h>
  #include <libqe/libqe.hpp>
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/epistemic-analytics/qe-packages/libqe}
  \item Report bugs at \url{https://gitlab.com/epistemic-analytics/qe-packages/libqe/-/issues}
}

}
\author{
\strong{Maintainer}: Cody L Marquart \email{cody.marquart@wisc.edu} (\href{https://orcid.org/0000-0002-3387-6792}{ORCID})

}
